Create table_pagination_h2o3.py
Based off of the `table_pagination_pandas.py` example but in theory this should work for any h2o-3 DataFrame with very minimal changes to the code. All changes end-users may want to make are marked with a `TODO`.

I am not sure if other files need to be updated when you make a new example.