fix: wave fetch failing because of incorrect url generation.
URL  should begenerated for fetching should be  of the format: `wave-{version}-{operating_system}-{arch}`

But the fetch url being generated was using `platform` instead of the `operating_sysyem`.

Since this variable is set on the build time. Changed the `hatch_build.py` file. This will also change the `metadata.py` file that gets generated.

Refs: h2oai#2177 (closes this issue)#